PN-10004, PN-10019

+++ PAS_Database/dbo/Stored Procedures/DeleteSubWorkOrderMaterialKit.sql
@@ -10,6 +10,7 @@ EXEC [DeleteSubWorkOrderMaterialKit]
 ** PR   Date			Author				Change Description  
 ** --   --------		-------				--------------------------------
 ** 1    12/12/2023	 Devendra Shekh			Created
+** 2    16/10/2024  RAJESH GAMI      Un Mapped PO by WO-SubWO Materials Id | KIT, While Delete the Materials
 
 exec dbo.[DeleteSubWorkOrderMaterialKit] 17
 **************************************************************/ 
@@ -23,6 +24,35 @@ BEGIN
 		BEGIN TRY
 			BEGIN TRANSACTION
 				BEGIN
+
+				IF OBJECT_ID(N'tempdb..##TempTableWOM') IS NOT NULL
+					BEGIN
+						DROP TABLE #TempTableWOM
+					END
+					CREATE TABLE #TempTableWOM(SubWorkOrderMaterialsKitMappingId BIGINT)
+					INSERT INTO #TempTableWOM (SubWorkOrderMaterialsKitMappingId)
+					SELECT SubWorkOrderMaterialsKitMappingId FROM [DBO].[SubWorkOrderMaterialsKitMapping] WHERE KitId = @KitId AND SubWOPartNoId = @SubWOPartNoId
+					
+					IF OBJECT_ID(N'tempdb..##TempWOtblM') IS NOT NULL
+					BEGIN
+						DROP TABLE #TempWOtblM
+					END
+
+					CREATE TABLE #TempWOtblM(SubWorkOrderMaterialsKitId BIGINT)
+					INSERT INTO #TempWOtblM (SubWorkOrderMaterialsKitId)
+					SELECT DISTINCT WOM.SubWorkOrderMaterialsKitId
+					FROM dbo.[SubWorkOrderMaterialsKit] WOM WITH(NOLOCK) INNER JOIN #TempTableWOM tmp ON WOM.SubWorkOrderMaterialsKitMappingId = tmp.SubWorkOrderMaterialsKitMappingId
+					WHERE WOM.SubWorkOrderMaterialsKitMappingId = tmp.SubWorkOrderMaterialsKitMappingId
+
+					UPDATE P    
+				    SET WorkOrderMaterialsId = 0, 
+					       IsKit = 0, IsSubWO =0, 
+						   UpdatedDate = GETUTCDATE()
+					FROM DBO.PurchaseOrderPart P
+					  INNER JOIN #TempWOtblM tmp ON P.WorkOrderMaterialsId = tmp.SubWorkOrderMaterialsKitId
+					  WHERE P.WorkOrderMaterialsId  = tmp.SubWorkOrderMaterialsKitId AND ISNULL(IsKit,0) = 1 AND ISNULL(IsSubWO,0) = 1
+
+
 					DELETE FROM [dbo].[SubWorkOrderMaterialStockLineKit] WHERE SubWorkOrderMaterialsKitId IN (SELECT SubWorkOrderMaterialsKitId FROM [DBO].[SubWorkOrderMaterialsKit] WHERE [SubWorkOrderMaterialsKitMappingId] IN (SELECT [SubWorkOrderMaterialsKitMappingId] FROM [DBO].[SubWorkOrderMaterialsKitMapping] WHERE KitId = @KitId AND SubWOPartNoId = @SubWOPartNoId));
 					DELETE FROM [DBO].[SubWorkOrderMaterialsKit] WHERE [SubWorkOrderMaterialsKitMappingId] IN (SELECT [SubWorkOrderMaterialsKitMappingId] FROM [DBO].[SubWorkOrderMaterialsKitMapping] WHERE KitId = @KitId AND SubWOPartNoId = @SubWOPartNoId);
 					DELETE FROM [DBO].[SubWorkOrderMaterialsKitMapping] WHERE KitId = @KitId AND SubWOPartNoId = @SubWOPartNoId;
